Add a file parser for parsing from the local file system.
When using Roboelectric (or other unit testing framework for Android that
does not set up a "real" Android environment), it's necessary to retrieve
files based on the local file system of the machine on which the tests are
run, rather than the virtual file system expected by the Android operating
system. This is due to the fact that, when running these types of tests,
most of the Android system is really just a set of mocks. As such, without
this file parser, the "assets", "res", and other system-level file system
paths are not set up and linked correctly, as the system calls that would
normally provide this information are mocked.